deck: scrollable PLACEMENT / GROUP / DATAFORTS lists; clear clippy
PLACEMENT (node page), GROUP (daemon page), and DATAFORTS (dataforts
tab) now window their contents against the cursor via a shared
`tabs::scroll_window` helper and surface "N more" indicators when
items are hidden above/below the viewport.

Clippy pass on the workspace:
- rpc-ffi `CancelEntry`: drop manual `Default` impl in favor of derive
- meshdb-ffi: drop unnecessary `unsafe` block on safe FFI calls
- bindings/node `common`: gate the module on its consumer features
  so the helpers stop reading as dead code under the `redis`-only
  default
- deck `App::new`: bundle the five tail handles into `streams::Tails`
  to drop under clippy's `too_many_arguments` ceiling
- deck `tabs/logs.rs`: delete unused duplicate `format_ts_ms`
- deck `tabs/blobs.rs`: drop unnecessary `u32` cast
- deck `tabs/net_map.rs`: introduce `GraphLayout` type alias
- sdk compute: `cfg_attr`-gate the groups-only fields / handle /
  method so they don't trip dead-code on non-groups builds

+++ net/crates/net/deck/src/tabs/blobs.rs
@@ -142,7 +142,7 @@ fn render_table(
             (None, None) => "—".to_string(),
         };
         let repl_style = match (e.replicas_observed, e.replica_target) {
-            (Some(o), Some(t)) if o < t as u32 => theme::amber(),
+            (Some(o), Some(t)) if o < t => theme::amber(),
             (Some(_), Some(_)) => theme::green(),
             _ => theme::dim(),
         };
